Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Dennis; hmmm - so Calc is increasingly an outlier here really ... we do want to do this for all components except Calc I think - and in future - we want to do it for calc too I think so ... =)

Hi Michael, I understand that we want to broadcast messages(independent of view zoom) in the core for all apps eventually, but I'm not sure why we want to broadcast the core broadcasted messages again in the kit. It would just blow up the message count. For example, in js console (at least in calc) I see a large number of duplicate messages (if we have many clients for a document) even for a single key press.

Patch-set: 1
Reviewer: Gerrit User 1000947 <1000947@6622c302-437a-49ad-a91d-55ba7dcf1615>