Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)

> Make sure you have an user called `lool` which can run /sbin/setcap with sudo without password. Make even sure the lool user owns all the directories used to build.

It is not required. You can 'make' with a normal user, then 'make install' as root. The lool user is only needed for running loolwsd, not for building it. 

I'd never install software made from source to /usr, rather I'd install it to /usr/local or to /opt. So thosre prefixes are fine for the docker image, but not in general.

For developers even that is not necessary, with --enable-debug configure option 'make run' works from the build tree, without setcap and lool user.

Patch-set: 2
Reviewer: Gerrit User 1000028 <1000028@6622c302-437a-49ad-a91d-55ba7dcf1615>