Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+2

As far as I can understand this script is used to generate a key but for what is this key used? Because if it's used for encryption, generating it once in the Dockerfile will expose the private key once the image is pushed to Docker HUB, internal registry, ...
If so, should be moved to the Docker entrypoint.

Patch-set: 3
CC: Gerrit User 1002429 <1002429@6622c302-437a-49ad-a91d-55ba7dcf1615>