Update patch set 2
Patch Set 2:

I think the cypress tests are failing because of the ID changes.
Where we use 'input#<ID>' selector in the test code, now we need to use something '#id input', I think. Are you aware that with removing this ID, the styling might change too? It might worth checking that this ID is not used in the CSS.

Patch-set: 2