Lets make 'dropdown marker' icon active
Before this, the dropdown marker icon was just trying to mirror the
exact shape and position of the validity list button in core.
It is a lot easier just to have the js icon according to our
styling and in its click handler, sent '.uno:DataSelect' command to
core. It saves us the trouble of mirroring the core's button which
apparently is not zoom adjusted. With non-default dpiscale, it takes
more effort to mirror it exactly.

Change-Id: Ibe15293e9052dfacf01be94ed4950d1842e46c4b