fix validation dropdown icon size
There is no need to have a 'maxHeight' for the icon as we don't mirror
the core's button anymore. The height of the icon follows the cell height,
and lets make the width follow it too, so that the icon is nicely
scaled with zoom (the icon used has a 1:1 aspect ratio).

Change-Id: I175ad98f6a1306fabd003eee7fda9652f5835995