wsd: reduce number of mounts
This reworks the approach used to bind-mount
the jails and the templates inside such that
the total is now down to only three mounts:
systemplate, lo, tmp.

Since now systemplate and lotemplate are
shared, they must be mounted as readonly,
this means that user/ must now be moved
into tmp/ which is writable.

The mount-points must be recursive, because
we mount lo/ within the mount-point of
systemplate (which is the root of the jail).
But because we (re)bind recursively, and
because both systemplate and lotemplate are
mounted for each jails, we need to make them
unbindable, so they wouldn't multiply the
mount-points for each jails (an explosive
growth!)

Change-Id: Iaddac9ab648de4d4835e0d1084ad04dae8c9a76e