Update patch set 3
Patch Set 3:

some suggestions (a quick analysis):

try to avoid adding "if doc layer", The annotation is a module (or class)
independent from document type, so some day it just can be replaced with
another implementation or for another platform. (but always there are some specific exceptions)

actually the edit paste issue, there are some global function that is listening the event
and when it receives it paste the text to document and prevent default action, it should not happen since the annotation edit got the focus. a hint for you is debug that global listener and analyses why is processing that event

Patch-set: 3