Update patch set 2
Patch Set 2:

Cypress tests are actually failing on this co-4-2-3 branch because of a crash during cypress verify command.
Check the log. For example this one:
https://cpci.cbg.collabora.co.uk:8080/job/gerrit_co-4-2-3_debug/167/console

The build was not failed because of the wrong error detection.
See this change:
https://gerrit.libreoffice.org/c/online/+/93917

Patch-set: 2
Reviewer: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>