Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Mert: I see that the code defines a isDarkModeActive variable, but it is only set and never read, is that intentional?

Indeed, I figured that we may need to send it to javascript or to the HTML even before load the document in order to trigger the dark theme for online. However, if it is not needed, i can update the patch?

Patch-set: 3
Reviewer: Gerrit User 1001306 <1001306@6622c302-437a-49ad-a91d-55ba7dcf1615>