cypress: fix running single unit test
When builddir != srcdir

Stills error:

"The support file is missing or invalid.

Your `supportFile` is set to `cypress_test/support/index.js`, but either
the file is missing or it's invalid. The `supportFile` must be a `.js`
or `.coffee` file or, if you're using a preprocessor plugin, it must be
supported by that plugin.

Correct your `cypress.json`, create the appropriate file, or set
`supportFile` to `false` if a support file is not necessary for your
project."

Change-Id: If1aeb55a3821735022dfa57af217130f18cd9dbf