Update patch set 1
Patch Set 1:

Why is this useful? Is it so hard to run 'make run-mobile' and then select the specific test suite? Is it easier to type the test suit name into the command line?
Maybe it would be more convenient to add a spec option for the 'make run-mobile' command too (and for make run-desktop).

Patch-set: 1
CC: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>