leaflet: restore reuse_cookies to the url
Reuse cookies are captured on file-serving the
loleaflet.html, whereupon they are injected
into it before serving as part of the document
load url. They show up as &reuse_cookies=
in the url embedded in loleaflet.html.
When loading the document, and creating
the session, the cookies received via
reuse_cookies is stored and set in each
WOPI http invocation.

This restores creating reuse_cookies in the
URL. It was inadvertently dropped after some cleanup.

Change-Id: I6de8ddb58e7f43566a08c22327dbc4b7a1207388