Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

I see, I think it does not merge the two options, but it's enough to specify the folder and it will search for index.js in the specified folder. It might be because of backward compatibility reasons. So I think you should still remove the same option from cypress.json to make it clear where we set that option. I guess the --config command-line option has the priority so the cypress.json does not have an effect anymore.
Also might be better to use ...support/index.js to make it more explicit. The related option is supportFile, which indicates a file and its default value is cypress/support/index.js, so it's meant to point to a file.
See also my inline comments.

Patch-set: 3
Reviewer: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1