leaflet: smart throttling of RichProxy polling
When there is no data coming in from WSD, we slow
down the frequency of invoking the RichProxy.

We don't invoke RichProxy more than 40 times a second
(i.e. with 25ms interval) and we slow down when
the last data was received 3x longer than the
polling interval.

However, when we want to send data, or when we do
receive any data, we immediately maximize the
frequency so we can communicate as fast as possible
and reduce the user-visible latency. Notice that
when we get some data (without having sent anything
recently) it implies activity from another user,
so we do want to get their changes in real-time.
This is why we reduce the polling frequency
gradually, but increase it abruptly.

The parameters of the algorithm are configurable
directly in the, but the current defaults are
very reasonable.

Change-Id: I0b9fd6db73caa35853fe6d8077bef66934de679c