Update patch set 6
Patch Set 6:

Currently, the support file includes only the cypress-failed-log module which is used in case of error to dump the log on the command line. So you can verify that this patch works if you break one of the tests and run make check-mobile or make check and check whether the log is dumped in the command line.

Patch-set: 6
Reviewer: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>