wsd: expose version-hash programmatically
Passing --version-hash to loolwsd now prints the
hash and exits. This is designed to be used with
proxy.php to check the version running.

Also, include the version and hash in the
hosting/capabilities JSON response.

Note that --version still prints/logs the version
and doesn't exit. It seems that is designed to
log it before running, and so it might be used
in the wild. Changing the behavior in minor upgrade
might break some production servers that have
--version added to their execution command-line.

Change-Id: I30e6945f797a902ed4b3aae6f5f1b8518000946b