Avoid requesting mobile wizard's content twice.
It's the JS code which triggers two events about
the sidebar content. We send a '.uno:SidebarShow'
and a 'dialogevent {"id":"-1"}' message. It does
not make sense to have both when opening the mobile
wizard.

The dialogevent behave more consistantly. It always
triggers the sidebar content, so it's better to use that
without a 400 ms delay.

With this fix we can remove the related workarounds.

Change-Id: I379b569264236a0729f775960607ecf8845de297