Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> Hello,
> 
> Thank you for your comment. I run the relevant command but i am not sure if it is working. It updated the npm-shrinkwrap.json file. I added the file to the patch manually in order to show the change. Also, shrinkwrap is not working with npm version > 3 according to their site.

You are welcome, You should run "shrinkpack": "1.0.0-alpha" that is installed in your local node_module folder, but it is OK, I can do the update later.

> 
> Below are some notes about the patch:
> 
> * In Writer, top menu is sending click commands twice. Calculator and impress is fine. When i click for example Format>>Paragprah button, page opens 2 dialogs. Do you have idea about this?

No, Thanks for your findings, if you can provide another patch to fix it, it will be great

> 
> Packages:
> 
> Dropped:
> @types/jquery: Since we are not using typescript, it is not needed.
> exorcist: Not used.
> 

If we are not using it should be removed, in any case in near future we can restore if it is needed it

> Not touched:
>     bootstrap: will be removed.

It is used in admin console, Please if you can analyze the admin Javascript files

>     eslint: too many errors: Eslint's new version is very sensitive and i think buggy.
> Last version counts below line as duplicate declaration:
> /* global $ */
> 

Better not to upgrade

> Replaced:
> uglify-js: Does not support es6 so it is replaced with terser.
>

We are not supporting ES6 for now, but at some point we will migrate all source to
ECMAScript 6 standard. Please just upgrade uglify-js
 
> Others:
> * Controlled (do we need this, what it does etc.).
> * Updated (of which have new versions).

No idea here, if you can provide link of the source code.

> 
> Changed:
> JQuery-ui: npm dist package does not contain template files. Lightness's template files are copied to outside of modules folder. They are copied to their original location with make command.
>

We are not importing files to images folder from jquery-ui, let me test it.
 
> Vulnerabilities:
> eslint: Can not update. It shows too many false positives.
> Bootstrap: Will be removed.
> shrinkpack: It does not support npm version > 3.
> 
> Some of the vulnerabilities are solved with running "npm audit --fix" command inside loleaflet folder.
>

Please add the command in the rule node_modules in Makefile.am

> makefile.am file is modified. Package needs below commands to be rerun, in the order (to be safe at least):
> 
> * inside loleaflet folder: npm install
> * autogen.sh
> * configure
> 
> Thank you.
> 
> Best Regards.

Patch-set: 5
Reviewer: Gerrit User 1000865 <1000865@6622c302-437a-49ad-a91d-55ba7dcf1615>