Update patch set 1
Patch Set 1:

We have a mention of select2@4.0.1 in loleaflet/Makefile.am, I assume we have to change the version there, too. 

But that leads to the question whether we should keep select2.js in loleaflef/js.
Henry introduced this change with https://gerrit.libreoffice.org/c/online/+/89110 but later we did no patch select2, so maybe it was unnecessary after all. Any thoughts, Henry?

Patch-set: 1