cypress: fix false failure of form field tests.
The cypress test framework checks if an item is covered
by an other item while we try to interact with it and
fails with an error in this case.
Since the form field button is above the document layer,
cypress correctly identifies that the document layer is
covered (partly) with this button frame. To avoid this
failure we can use force : true option.

Change-Id: I5d9f665e3afe1ae1ff3744a563dd32de52b76bb3