wsd: leaflet: fix reuse_cookies support
reuse_cookies is now always encoded in the URL.

And, there is no need for the WOPISrc in the three cases
in this patch, and by passing the DocumentURI proper
(without /ws?WOPISrc=...) ensures that all query-params
in the DocumentURI are properly processed.

This fixes the reuse_cookies regression where it
wasn't passed to WOPI requests.

Change-Id: I8dccfb09a7b4102d10c1aef24f43b699a07bfed8