cypress: better way of filter out tests based on core version.
We use cypress-select-tests to filter out tests before
the tests are started. One advantage is that these tests
are showed as skipped in the log and also beforeAll() is
not called for these test cases which can speed things up.

Also we can avoid to add too much noise to the test code.
When we are working with more branches, we can just add
blacklists separately for all branches.

Change-Id: Ie7808614e42a19a0820f372720cc391511e165f5