leaflet: throttle requests to avoid overwhelming the server
Previously the proxy requests were made as fast as possible,
which was 25ms, regardless of how responsive the server is.

Now we are careful to throttle when the server returns
and error before we retry, and we take into account the
roundtrip of requests so we don't make requests faster
than half the roundtrip.

Also, handle connection timeouts and throttle when
they happen.

Formatting to make the braces consistent.

Change-Id: Icd8fe7b00ed7e9ca53c61adfe99ebbc6e07757f4