Update patch set 5
Patch Set 5:

I think we should not rely on the IDE(vscode) invoking tsc via compileOnSave option as it breaks the non-interactive builds (in this case, Jenkins would not have generated js for the new ts file introduced in this patch, and there are no tests to catch load errors, at least in admin/). Instead lets invoke it explicitly in Makefile(.am) under a new rule for .ts -> .js, so that the top-level 'make' / 'make run' will ensure all ts files are converted.

Patch-set: 5
CC: Gerrit User 1000947 <1000947@6622c302-437a-49ad-a91d-55ba7dcf1615>