Update patch set 1
Patch Set 1: Code-Review+2

Looks really good =)

I'd like a LOG_ERR() when we have to kill a process for not responding - it is really an error in the S/W somewhere IMHO and many installations have only error logging.

I'd also love to get a stack trace from that process - if we kill with SIGABRT first and give it a few seconds to dump a trace, then move up to SIGKILL that might provide us with useful data: why (and where) are these processess hanging.

But perhaps that's for a follow-on patch (?)

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+2