Update patch set 3
Patch Set 3:

Some Cypress tests still fail on this patch.
There was a bug in the cypress test framework, that's why Jenkins show it as passed, but if you go through the log you can find "Timed out retrying" errors.
Let me rebase this patch to fix this issue and to see which tests are failing on this.

Patch-set: 3
CC: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>