cypress: test download as options.
We can't test the actual download:
https://github.com/cypress-io/cypress/issues/949

What we test is that we get a proper URL, after triggered
the download as menu item. This src attribute is set after
the core exported the file and sent back the URL for it.

We avoid download for test, because we can't control the
download folder. It's better not to save files without
clean up. Downloading is not working in headless mode
anyway.

Change-Id: I93664be7f55e4df863448adc734fa9cfc7b3638f