Update patch set 17
Patch Set 17:

(I mean, the number of threads per document is the same as before. If there is just one, the number will be the same as before. If there are multiple documents, there will be additional threads sure, but the alternative would be heavy re-engineering.)

Patch-set: 17
Reviewer: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>