fall back to en_US.UTF-8 if C.UTF-8 does not exist
It turned out that C.UTF-8 is not supported on RH7/CentOS7.
We started to use C.UTF-8 because it was widely available and we
could shave off locale data in Ubuntu based docker images and
AppImage. With this patch we fall back en_US.UTF-8 if C.UTF-8 does
not exist and add some logging.

Change-Id: Idea0ae885dc8cdd9ef33279bd90eb882a656d75c