print-twips messaging for graphics
All graphic-selection messages from core are in print-twips if
sc_print_twips_msgs is set in SAL_LOK_OPTIONS. So do all the necessary
conversions on intercepting the message.

In addition, in this mode the core expects the parameters of the command
'.uno:TransformDialog' to be in print-twips. This is also done in this
patch.

Change-Id: Ief8a0d4677f83f26b135ce47b68c27bff5aaf5c9