Wrong "perm" value when setting permission, works by accident
When "perm:" message is sent, in case of PDF we send the payload
"perm: readonly", which is then parsed as " readonly" (with extra
space). If a proper "readonly" value would be parsed, JS would
get stuck, because the code assumes that "doclayer" is available,
which is not the case. So this fixes that the command is correctly
parsed and that it doesn't get stuck by not running the code that
assumes doclayer is available.

Change-Id: I35b6cc25209b4ed259f33f7fb77bc0be7a69033e