wsd: admin: don't poll rapidly when cleanup is disabled
When per_document.cleanup is disabled, the time
between the last cleanup (which never happened)
grows indefinitely, which results in minimal
polling time intervals. This wastes valuable
cpu cycles unnecessarily.

When cleanup is disabled, there is no need to
calculate the next cleanup time. The maximum
is reasonable (although it should really be
infinity).

Change-Id: I71d065441c4c2ff96fe31e6a45a5ecfdd2f85d49