Update patch set 4
Patch Set 4:

New patch at https://gerrit.libreoffice.org/c/online/+/97699

My intention with this patch is to provide feedback to the integrator about the save result (with Action_Save_Resp).
So this does not emit a notice that is visible to users, but only to the integrator. And the integrator only sees the notification if he requested to see them by setting Notify=True (in Action_Save).
And he will only get one notification for the save action he requested. After that _notifySave in Map.WOPI.js is reset to false and no further notifications will be sent to the integrator.

So for that to work correctly, I need to return the save result in every case (before each return statement).

Does that make sense? If you think this can be improved somehow, let me know.
Maybe the term broadcastSaveResult is a bit confusing here and should rather be emitSaveResult or something?

Patch-set: 4