Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hmm - we -used- to have a different random sub-path for jails for each test; did that disappear ? I would strongly suggest restoring that if so. Otherwise we have to really weaken our "cleanup where the jails are" code that runs before and afterwards to avoid races.

We do have random sub-paths. This is to update the "dynamic" files in systemplate (etc/passwd, etc/hosts, ...).

The patch is incomplete. But we need a safe way to update these files during tests. In production, this isn't an issue, as there is only one loolwsd and and forkit. During testing we have many, and they all need to update systemplate, which causes a race that breaks the tests.

Patch-set: 1
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>