android: On ChromeOS, load files from Google Drive read-only.
Apparently it is not possible to save to Google Drive yet.
Unfortunately I don't have a proof in the documentation or anywhere on
the net, but the good indication of this is that the
Intent.ACTION_CREATE_DOCUMENT does not offer Google Drive (while it does
on a "normal" Android).

Change-Id: I521b866fd783b81a2bb9eace84354b5564ebabee