Update patch set 12
Patch Set 12:

(2 comments)

As I see, there is a lot of flickering on this dialog. I think now it's the best to reduce this test case to search only once and that's all, so we can avoid false failures. We can test the search dialog later in a separate test file.

Patch-set: 12
CC: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>