cypress: introduce clickOnIdle() and inputOnIdle() methods.
This method waits until the item is idle,
e.g. not detached for a while. Using this we can
workaround false failures caused by GUI flickering.
For example, mobile wizard is updated all the time
which makes hard to test it reliably. We can use
this clickOnIdle() method, which is slower than the
simple click(), but is more reliable.

Change-Id: I2f970eb0cf400382c8384c91ab7c84b1e02e63af