Fix create-l10n-all-js.pl for languages that include script and territory
On Android and iOS, the window.LANG gets set to a value that uses
hyphens as separators, not underscores. Also, the script is included
for the Chinese variants. While at it, make the code more lenient:
Accept either separator and with script or not.

Change-Id: I5a5319572ec96892d22eb3b30586b5b8203bdee1
Reviewed-on: https://gerrit.libreoffice.org/c/online/+/98434
Tested-by: Jenkins
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tor Lillqvist <tml@collabora.com>
(cherry picked from commit 8166a0edaf93ab030af68ddac52910d7a93a2e0a)