loleaflet: fix nodocloaded error thrown while reconnecting to a xlsx file
The map._activate, among other actions, is sending indirectly some messages
to the server like clientzoom and clientvisiblearea. If these messages are send
before the document finishes processing the load message then there is
a chance that a nodocloaded error will be thrown because there is a
chance that the messages will be processed in parallel with load. This happens
constantly for xlsx files. This is generated by the Unipoll mechanism which,
in case of xlsx files, triggers a parallel processing.
To avoid the above scenario the call to _activate right after sending load
message was removed. It is done anyway after receiving the docloaded message
from server.

Change-Id: I4c83e72e600f92d0bb4f1f18cebe694e326256d0