Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Ooh - interesting =) nice catch. I guess these could enter through a progress bar callback which perhaps spins the main-loop to refresh the screen.

The parallel processing is triggered in core when the document's sheets are loaded: https://git.libreoffice.org/core/+/refs/heads/master/sc/source/filter/oox/workbookfragment.cxx#336. While waiting for worker threads to finish the main thread is doing some message processing in parallel, probably as an optimization made not to block the application. The weird thing here is that this is made only for LibreOffice own format types. BTW, do you know why this is not made for all table file types ?

> I wonder if (as a more robust idea) we could block incoming event processing during an un-completed load inside the online/Kit core.

I thought already about the solution you propose before proposing my current solution but I dismissed it as a more dramatic one because it means to block processing of other messages while loading and I'm not aware of the consequences even if we'll block only for Online. I thought that the parallel processing on load was implemented with something in mind and it shouldn't be dismissed. But if someone can guarantee that it's ok to do it then I can do it.

On the other hand, even if we'll implement the more robust idea should we keep my current proposal ? Are there any problems with it ?

Thanks

Patch-set: 1