Update patch set 1
Patch Set 1:

This is quite an 'exciting' piece of code. The problem is that the codebase in many, many places requires the "one big lock", while of course we can do a lot of work without taking it. So in order to let others get that lock needed in order to report progress and do various operations to the UI and the calc core - we do need to Yield.

But we could easily block this happening and input event processing during a load in the Kit code I think; it would be worth doing to make things more deterministic. Having had a poke around doing that by just buffering incoming data in the inBuffer on the socket more easily, in handlePoll:
        // If we have data, allow the app to consume.
By adding a "setQueueInput(bool )" to that socket - and just not processing input when we we are doing some bigger operations, and of course re-start processing when that is off.

Is that something you could look at ? =) I guess it'd make sense for Save as well as Load - to have a guard around them that sets all incoming sockets to queuing while they complete - I imagine that would nail a slew of un-detected and un-common re-entrancy hazards =)
 
If not, I'll try to get to it =)

Thanks !

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>