Update patch set 1
Patch Set 1:

Hello,

In calculator, there is an indicator at the bottom-middle of the page. When user selects more than one cells, the indicator show the selected row and column count. But its alignment is buggy.

This patch resolves the above mentioned bug. This is a workaround. Because it is very difficult to find the CSS rule causing the bug.

To review you may go to line 301 directly. Other changes are all removal of whites spaces.

May you review the patch.

Thank you.

Patch-set: 1
Reviewer: Gerrit User 1000865 <1000865@6622c302-437a-49ad-a91d-55ba7dcf1615>
CC: Gerrit User 1000028 <1000028@6622c302-437a-49ad-a91d-55ba7dcf1615>
CC: Gerrit User 1000863 <1000863@6622c302-437a-49ad-a91d-55ba7dcf1615>