Update patch set 2
Patch Set 2:

Do I understand it correctly, that this test do not actually test the use case you intended to test? The commit message indicates something like this.

The cypress test framework's mobile emulation is imperfect so it's not suprising that something is not working on the same way as the actual application. Is this a mobile only issue? If not a desktop test might be a better choice.

Otherwise you should not add a test which is just a placeholder. The test framework should be improved in the future and after that this use case also can be covered with a test case.

Patch-set: 2
CC: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>