Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Specifically, after the patch, the "1970-01-01" date is not shown, but the words "Last modification" are visible.

hm odd, when I type in the browser console "$('#menu-last-mod').hide()" the whole thing seems disappears (via display: none) and it works. So if this is not happening when it arrives to that condition then maybe that label gets created again.

I also wonder if maybe mixing L.DomUtil and jquery might not be ideal so maybe it would be better to just stick with one.

So either 100% L.DomUtil:
"L.DomUtil.setOpacity(lastModButton, 0);"

or

jquery right from the moment of getting the object:
"var lastModButton = $('#menu-last-mod');"

Patch-set: 2
Reviewer: Gerrit User 1002126 <1002126@6622c302-437a-49ad-a91d-55ba7dcf1615>