Update patch set 1
Patch Set 1: Code-Review-1

Hi Szymon, this may work around it, but I'm not sure it's the right fix. It also gives us a rather busy-loop I think. The ideal is to have that poll timeout be ~infinite when there is no work to do. We should know that there are queued events to process, and then set the timeout to zero in this case (prolly we still want to poll) - but we're not doing that properly for some reason. Can you poke again ? =)

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1