Update patch set 1
Patch Set 1:

I did profiling runs with and without this, where I attempted to do the same: Load the same document, zoom in to the same spot and magnification, and close. With the png encoding the app process consued 9.3 seconds CPU (with the png encoding indeed taking most of it). With the patch, it took 6.6 seconds. I don't know how much CPU the WebKit process took.

Patch-set: 1