android: Add support for ODT/ODS/ODP for Chrome OS.
Turns out that the ODF mime types are most probably not supported by
Chrome OS, because the android:mimeType=... is not enough to associate
the app with the ODT/ODP/ODS files.

So introduce an additional intent-filter that explicitly matches
according to the file extension; a bit hacky, but works.

Change-Id: If7cad42333314e8d16f96d035380460d5c3da347