tdf#129296 leaflet: Password protected document load failing
Do not send faileddocloading error if it is because of password

When the document is password protected, we send to the client both passwordrequired and faileddocloading.
These two are handled differently. While the first prompts the user for password input, the second
internally flags fatal error and shows an error message that the document may be corrupted etc.
The end result is that the document is not loaded and displayed when the user submits a correct password.
To reset the fatal error one has to reload, which is unhelpful when we need to provide a password.

This patch makes sure that we only send one error message to the client. If a password is required,
it already implies that the document didn't load, and that with the proper password we should try again.
Similarly for when the password given is wrong. However, if loading fails and it isn't a password-related
failure, faileddocloading error is returned and in this case the client handles
it as a final error (that requires reloading to retry).

Change-Id: I383418fd40b6e0749b20af0ef8dc40f391a05559