Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

So in general we have the cypress test framework, but it's not complete. Not all things can be tested easily in the current form. The sidebar is one of this area, beacuse it's just an image, we can't interact with the widgets on it directly. We should not use fixed positions, because core changes frequently and it would take too much time to update these tests. Just imagine 100 of this kind of tests and imagine the pain to fix all of them after a new item was added to the sidebar or the positon was just changed a bit in the css. We need to find a way to test the sidebar effectively or maybe later we will have a JS based sidebar similar to the one on mobile and so we will be in a better situation.

Patch-set: 1
Reviewer: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1