add AnnotationManagerBase to share code with other modules
Currently AnnotationManager is meant for Writer only, but it makes
sense to share the code with other modules too. This adds a new
base class AnnotationManagerBase which will contain the common
code for the annotation management. Other modules currently have
the annotations implemented in TileLayer, but this is the first
step to move them into its own annotation managers too.

Change-Id: I5131337aa95e43ce0844d57e2cc020239af5f100