cypress: add an option to dump only console.error() messages.
With the ENABLE_LOGGING flag we get all errors coming from
the browser, which is a lot. So I introduce ENABLE_CONSOLE_LOG
flag now, which dumps only console.error() messages, so we
can debug the client code easier when we run a test in headless mode.

It will be useful for multi-user tests, where we can't use the
interactive test runner, because we need to run more users at
the same time.

Change-Id: Icfa62412b83e37398c54e9ac7b4120d76a87fb92